commit bitcoin get currency
