+++ bitcoin-api/src/main/java/coindesk/model/DataCurrency.java
@@ -1,0 +1,23 @@
+package coindesk.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class DataCurrency {
+    private String code;
+    private String symbol;
+    private String rate;
+
+    public String getCode() {
+        return code;
+    }
+
+    public String getSymbol() {
+        return symbol;
+    }
+
+    public String getRate() {
+        return rate;
+    }
+
+}
