+++ bitcoin-api/src/main/java/coindesk/model/TimeBuilder.java
@@ -1,12 +1,13 @@
 package coindesk.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class TimeBuilder {
+
+    @JsonProperty("updated")
     private String updated;
 
-    public TimeBuilder(){
-
-    }
-    public TimeBuilder(String updated) {
-        this.updated = updated;
-    }
 }
