--- esgi-tools/src/main/java/tools/Device.java
@@ -1,5 +1,0 @@
-package tools;
-
-public enum Device {
-    //Create all device Euro etc
-}
